LayoutTests/fast/harness/uiscriptcontroller/ui-side-script-with-callback-expected.txt

Test that a second immediate UI script returns if an earlier script registered a callback

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scriptResult is "async task complete"
PASS scriptResult is "immediate script complete"
PASS successfullyParsed is true

TEST COMPLETE
